requirements.txt
django-bootstrap5==24.3
django-cleanup==8.1.0
Pillow==10.0.0
django-ckeditor-5==0.2.15